% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{agr}
\alias{agr}
\title{Compute agreement between compartment calls}
\usage{
agr(x, ...)
}
\arguments{
\item{x}{A MultiCompartmentCall object}
}
\description{
Calculates the proportion of calls with the same sign for every pair of
calls in a MultiCompartmentCall object.
}
