% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompartments.R
\docType{data}
\name{array.data.chr14}
\alias{array.data.chr14}
\title{Example Illumina 450k methylation array data for compartmap}
\usage{
data(array.data.chr14, package = "compartmap")
}
\description{
This data was generated using the data from the reference via the
sesamize function from the SeSAMe package.
}
\references{
\url{https://f1000research.com/articles/5-1281/v3}
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
