% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCompartments.R
\name{bootstrapCompartments}
\alias{bootstrapCompartments}
\title{Non-parametric bootstrapping of compartments and summarization of
bootstraps/compute confidence intervals}
\usage{
bootstrapCompartments(
  obj,
  original.obj,
  BPPARAM,
  bootstrap.samples = 1000,
  chr = "chr14",
  group = FALSE,
  assay = c("rna", "atac", "array"),
  targets = NULL,
  res = 1000000,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  q = 0.95,
  svd = NULL,
  bootstrap.means = NULL
)
}
\arguments{
\item{obj}{List object of computed compartments for a sample with 'cscore' and 'gr' as elements}

\item{original.obj}{The original, full input SummarizedExperiment of all samples/cells}

\item{BPPARAM}{BiocParallelParam for parallelizing bootstrapping}

\item{bootstrap.samples}{How many bootstraps to run}

\item{chr}{Which chromosome to operate on}

\item{group}{Whether this is for group-level inference}

\item{assay}{What sort of assay are we working on}

\item{targets}{Targets to shrink towards}

\item{res}{The compartment resolution}

\item{genome}{What genome are we working on}

\item{q}{What sort of confidence intervals are we computing (e.g. 0.95 for 95 percentCI)}

\item{svd}{The original compartment calls as a GRanges object}

\item{bootstrap.means}{Pre-computed bootstrap means matrix}
}
\value{
Compartment estimates with summarized bootstraps and confidence intervals
}
\description{
Non-parametric bootstrapping of compartments and summarization of
bootstraps/compute confidence intervals
}
\examples{

# this needs a good example

}
