% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{bpnworkers.list}
\alias{bpnworkers.list}
\title{Return the number of workers in a list of BiocParallelParam objects}
\usage{
bpnworkers.list(bplist)
}
\arguments{
\item{List}{of BiocParallelParam objects}
}
\value{
A vector of the \code{bpnworkers} count in each list element
}
\description{
Return the number of workers in a list of BiocParallelParam objects
}
\keyword{internal}
