% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{check_optim}
\alias{check_optim}
\title{Check and tell users whether input BPPARAM are optimal given group/sc
inference and bootstrapping}
\usage{
check_optim(workers, group, chr_count, bootstrap)
}
\description{
Check and tell users whether input BPPARAM are optimal given group/sc
inference and bootstrapping
}
\keyword{internal}
