% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{check_worker_count}
\alias{check_worker_count}
\title{Check that the number of requested workers is valid}
\usage{
check_worker_count(
  bpparam,
  group,
  chr_count,
  bootstrap,
  avail_workers = parallelly::availableCores()
)
}
\description{
Check that the number of requested workers is valid
}
\keyword{internal}
