% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkAssayType}
\alias{checkAssayType}
\title{Check if the assay is a SummarizedExperiment}
\usage{
checkAssayType(obj)
}
\arguments{
\item{obj}{Input object}
}
\value{
Boolean
}
\description{
Check if the assay is a SummarizedExperiment
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
checkAssayType(array.data.chr14)
}
