% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanAssay}
\alias{cleanAssay}
\title{Generate function to filter rows/columns with NAs exceeding a threshold}
\usage{
cleanAssay(by = c("row", "col"))
}
\arguments{
\item{by}{Whether to filter by rows or columns}
}
\value{
A function to filter assay rows/columns
}
\description{
Generate function to filter rows/columns with NAs exceeding a threshold
}
\details{
Since removing NAs from rows vs columns only differs by whether rowMeans or
colMeans is used, and by where the comma goes in the subset operation,
code repetition can be avoided by consolidating these operations.
This \code{cleanAssay} function can generate two functions to remove NA's from
rows and columns using the \code{by} argument based on which it selects the
appropriate 'mean' and subset functions. This maintains the clarity of
having the operation in the function name when used: \code{cleanAssayRows} and
\code{cleanAssayCols}.
}
\keyword{internal}
