% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{CompartmapCall}
\alias{CompartmapCall}
\title{CompartmapCall class (experimental)}
\usage{
CompartmapCall(gr, res, name = NULL, unitarized = FALSE)
}
\arguments{
\item{gr}{The GRanges output of scCompartments or getArrayCompartments
containing the 'pc' column}

\item{res}{The binning resolution used}

\item{name}{An identifier for this compartment call. Important to set if you
are making a MultiCompartmentCall.}

\item{unitarized}{Whether the singular values have been unitarized}
}
\description{
An S7 class to hold a single-sample or grouped compartment call with its
metadata and analysis methods. This can take the GRanges output of
\code{scCompartments(group = TRUE)}. For multiple single-cell level compartment
inferences use MultiCompartmentCall.
}
