% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{compartments}
\alias{compartments}
\title{Estimate A/B compartments from methylation array data}
\usage{
compartments(object, resolution = 1e+06, what = "OpenSea", chr = "chr22",
  method = c("pearson", "spearman"), keep = TRUE)
}
\arguments{
\item{object}{}

\item{resolution}{}

\item{what}{}

\item{chr}{}

\item{method}{}

\item{keep}{}
}
\description{
\code{compartments} returns estimated A/B compartments from methylation array data
}
\examples{

}
