% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{compute_mahalanobis}
\alias{compute_mahalanobis}
\title{Get the Mahalanobis distance for the singular values}
\usage{
compute_mahalanobis(mat, cov_method = c("base", "robust", "mcd"))
}
\arguments{
\item{mat}{Matrix of singular values}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},
"robust": \code{robust::covRob()}, "mcd": \code{robust::covRob(estim = "mcd")}}
}
\description{
Get the Mahalanobis distance for the singular values
}
