% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseRE.R
\name{condenseSE}
\alias{condenseSE}
\title{Condense a the output of condenseRE to reconstruct per-sample GRanges objects to plot}
\usage{
condenseSE(obj, sample.name = NULL)
}
\arguments{
\item{obj}{Output of condenseRE or can be a RaggedExperiment}

\item{sample.name}{Vector of samples/cells to extract}
}
\value{
GRanges or list of per-sample GRanges to pass to plotAB or export
}
\description{
Condense a the output of condenseRE to reconstruct per-sample GRanges objects to plot
}
\examples{
grl <- GRangesList(GRanges(c("A:1-5", "A:4-6", "A:10-15"), score=1:3),
GRanges(c("A:1-5", "B:1-3"), score=4:5))
names(grl) <- c("A", "B")                  
x <- RaggedExperiment(grl)
condense.x <- condenseSE(x, sample.name = "A")
}
