% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{corr}
\alias{corr}
\title{Compute correlation between compartment calls}
\usage{
corr(x, ...)
}
\arguments{
\item{x}{A MultiCompartmentCall object}
}
\description{
Calculates Pearson correlation for every pair of calls in a
MultiCompartmentCall object.
}
