% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{dc_bins}
\alias{dc_bins}
\title{Get bins with significant Mahalanobis distances}
\usage{
dc_bins(gr, md, alpha_level = 0.05)
}
\arguments{
\item{gr}{Matrix of singular values}

\item{md}{data.table output from \code{compute_mahalanobis()}}

\item{alpha_level}{Significance level to use (default: 0.05)}
}
\description{
Get bins with significant Mahalanobis distances
}
