% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{DF}
\alias{DF}
\title{Get the \verb{@df} slot from a CompartmentCall object.}
\usage{
DF(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
\code{n}: bin indices corresponding to indices of the \code{GRanges} object in \verb{@gr}
\code{cscore}: compartment call singular values
\code{name}: The name of the individual \code{CompartmentCall} in a
\code{MultiCompartmapCall} object
}
\concept{s7getters}
