% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{differentiate}
\alias{differentiate}
\title{Compute the direction of change of singular vectors in the MultiCompartmentCall object}
\usage{
differentiate(x)
}
\arguments{
\item{x}{A \code{MultiCompartmentCall} object}
}
\description{
The direction of change is computed by chromosome
}
