% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{dim,compartmap::CompartmentCall-method}
\alias{dim,compartmap::CompartmentCall-method}
\title{Get dimensions of \code{MultiCompartmapCall} matrix}
\usage{
## S7 method for class <compartmap::CompartmentCall>
dim(x)
}
\description{
Get dimensions of \code{MultiCompartmapCall} matrix
}
\concept{s7getters}
