% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{dim,compartmap::MultiCompartmapCall-method}
\alias{dim,compartmap::MultiCompartmapCall-method}
\title{Get dimensions of \code{MultiCompartmapCall} matrix}
\usage{
## S7 method for class <compartmap::MultiCompartmapCall>
dim(x)
}
\description{
Get dimensions of \code{MultiCompartmapCall} matrix
}
\concept{s7getters}
