% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{-.compartmap::CompartmentCall}
\alias{-.compartmap::CompartmentCall}
\title{Get the difference between two CompartmentCall objects call values}
\usage{
\method{-}{`compartmap::CompartmentCall`}(x, y)
}
\arguments{
\item{x, }{y CompartmentCall objects to compare}
}
\description{
Get the difference between two CompartmentCall objects call values
}
