% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVD.R
\name{.getUV}
\alias{.getUV}
\title{Run SVD and get left or right singular vectors}
\usage{
.getUV(matrix, k, sing.vec)
}
\arguments{
\item{matrix}{A p x n input matrix}

\item{k}{Number of singular vectors to return}

\item{sing.vec}{Whether to return the right or left singular vector}
}
\value{
left or right singular vectors
}
\description{
Run SVD and get left or right singular vectors
}
\keyword{internal}
