% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkBins.R
\name{.jse}
\alias{.jse}
\title{James-Stein estimator}
\usage{
.jse(x, grand.mean = NULL, targets = NULL)
}
\arguments{
\item{x}{input vector of binned means across samples}

\item{grand.mean}{The global mean across samples}

\item{targets}{Samples to shrink towards

\eqn{\hat{\theta}_{JS+} = \left(1 - \frac{(m - 3)\sigma^2}{||\textbf{y} - \nu||^2}\right)}}
}
\description{
James-Stein estimator
}
\keyword{internal}
