% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrestiCoullCI.R
\name{.n_approx}
\alias{.n_approx}
\title{n_tilde in AC}
\usage{
.n_approx(n1, n0, q)
}
\arguments{
\item{n1}{number of successes/ones}

\item{n0}{number of failures/zeroes}

\item{q}{quantile for eventual CI (e.g. 0.95 for a 95\\% binomial CI)}
}
\value{
the "effective" sample size for smoothed CIs
}
\description{
n_tilde in AC
}
