% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrestiCoullCI.R
\name{.p_approx}
\alias{.p_approx}
\title{p_tilde in AC}
\usage{
.p_approx(n1, n0, q)
}
\arguments{
\item{n1}{number of successes/ones}

\item{n0}{number of failures/zeroes}

\item{q}{quantile for eventual CI (e.g. 0.95 for a 95 percent binomial CI)}
}
\value{
the approximate success probability for a smoothed CIs
}
\description{
p_tilde in AC
}
\details{
\eqn{\tilde{p} = \frac{1}{\tilde{n}}(n_{\text{success}} + \frac{z^2_\alpha}{2})}
}
