% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrestiCoullCI.R
\name{.z}
\alias{.z}
\title{Normal alpha/2 quantile}
\usage{
.z(q)
}
\arguments{
\item{q}{the quantile at which to extract Z}
}
\value{
Z
}
\description{
Normal alpha/2 quantile
}
\details{
\eqn{z_\alpha = \Phi^{-1}(1 - \frac{\alpha}{2})}
}
\keyword{internal}
