% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingPrep.R
\name{embeddingPrep}
\alias{embeddingPrep}
\title{Filter and convert/cluster compartment calls for embedding and dimension reduction}
\usage{
embeddingPrep(
  obj,
  filter = TRUE,
  min.conf = 0.7,
  min.eigen = 0.02,
  tf.idf = FALSE,
  cluster = TRUE,
  k = 10,
  dims = NULL
)
}
\arguments{
\item{obj}{Input either a RaggedExperiment or output from fixCompartments}

\item{filter}{Whether to filter compartments}

\item{min.conf}{Minimum confidence estimate to use when filtering}

\item{min.eigen}{Minimum absolute eigenvalue to use when filtering}

\item{tf.idf}{Whether to TF-IDF transform the data}

\item{cluster}{Whether to identify clusters in the data}

\item{k}{How many neighbors when identifying clusters}

\item{dims}{How many dimensions to use for identifying clusters}
}
\value{
A prepared matrix of list of compartment calls for embedding and dimension reduction
}
\description{
Filter and convert/cluster compartment calls for embedding and dimension reduction
}
\examples{

}
