% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractOpenClosed}
\alias{extractOpenClosed}
\title{Get the open and closed compartment calls based on sign of singular values}
\usage{
extractOpenClosed(gr, cutoff = 0, assay = c("rna", "atac", "array"))
}
\arguments{
\item{gr}{Input GRanges with associated mcols that represent singular values}

\item{cutoff}{Threshold to define open and closed states}

\item{assay}{The type of assay we are working with}
}
\value{
A vector of binary/categorical compartment states
}
\description{
Get the open and closed compartment calls based on sign of singular values
}
\examples{

dummy <- matrix(rnorm(10000), ncol = 25)
sing_vec <- getSVD(dummy, k = 1, sing.vec = "right")

}
