% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndImputeWGBS.R
\name{filterAndImputeWGBS}
\alias{filterAndImputeWGBS}
\title{Filter and impute missing values prior to A/B compartment inference with k-nearest neighbor}
\usage{
filterAndImputeWGBS(obj, k = 0.5, minCov = 3, minSamp = 2, rowmax = 0.5,
  colmax = 0.8, openseafilt = FALSE, genome = "hg19")
}
\arguments{
\item{obj}{Input BSseq object}

\item{k}{The pseudocount to be added for smoothing - default uses Jeffery's prior}

\item{minCov}{Minimum coverage of a locus to be smoothed}

\item{minSamp}{Minimum samples that have sufficient coverage to be smoothed}

\item{rowmax}{The maximum percent missing data allowed per loci - default = 0.5}

\item{colmax}{The maximum percent missing data allowed per sample - default = 0.8}

\item{openseafilt}{Whether to filter loci to just open sea CpG loci}

\item{genome}{The genome to analyze - currently only supports hg19}
}
\value{
Filtered and imputed matrix
}
\description{
\code{filterAndImputeWGBS} returns filtered and smoothed signal from whole genome bisulfite sequencing data
}
\details{
This function filters and performs Dirichlet smoothing of WGBS signal prior to A/B compartment inference
}
\examples{

#Load in some example data
data(cell_cycle_hansen_chr14.rda, package = "compartmap")

#Filter and impute missing data
filt.imputed.chr14 <- filterAndImputeWGBS(data.chr14)
}
