% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells using either k-means or Dirichlet process means clustering of sparsity metrics}
\usage{
filterCells(sparsity.mat, rse.obj, cluster.method = c("kmeans",
  "dpmeans"), clusters = 2, plot.data = FALSE)
}
\arguments{
\item{sparsity.mat}{A matrix of summarized sparsity measures}

\item{rse.obj}{The unfiltered RangedSummarizedExperiment object}

\item{cluster.method}{Clustering method to use (default: kmeans)}

\item{clusters}{How many clusters to generate}

\item{plot.data}{Whether to plot the data}
}
\value{
A filtered RangedSummarizedExperiment object and/or plot of the filtered data
}
\description{
Filter cells using either k-means or Dirichlet process means clustering of sparsity metrics
}
\examples{

}
