% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCompartments.R
\name{filterCompartments}
\alias{filterCompartments}
\title{Filter compartments using confidence estimates and eigenvalue thresholds}
\usage{
filterCompartments(obj, min.conf = 0.7, min.eigen = 0.02)
}
\arguments{
\item{obj}{Output of condenseSE or fixCompartments}

\item{min.conf}{Minimum confidence estimate to use when filtering}

\item{min.eigen}{Minimum absolute eigenvalue to use when filtering}
}
\value{
A filtered/subset of the input object/list
}
\description{
Filter compartments using confidence estimates and eigenvalue thresholds
}
\examples{

}
