% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompartments.R
\docType{data}
\name{filtered.data.chr14}
\alias{filtered.data.chr14}
\title{Example ATAC-seq data for compartmap}
\usage{
data(filtered.data.chr14, package = "compartmap")
}
\description{
This data was generated using the data from the reference via bwa mem
and pre-processing the data using the ATACseeker package.
}
\references{
\url{https://trace.ncbi.nlm.nih.gov/Traces/sra/?study=SRP082417}
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
