% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterOpenSea}
\alias{filterOpenSea}
\title{Filter to open sea CpG loci}
\usage{
filterOpenSea(obj, genome = c("hg19", "hg38", "mm10", "mm9"), other = NULL)
}
\arguments{
\item{obj}{Input SummarizedExperiment or GRanges object}

\item{genome}{Which genome to filter}

\item{other}{GRanges of open sea regions (TODO)}
}
\value{
Filtered to open sea CpG loci
}
\description{
Filter to open sea CpG loci
}
\examples{
if (requireNamespace("minfi", quietly = TRUE)) {
  data("array_data_chr14", package = "compartmap")
  opensea <- filterOpenSea(array.data.chr14, genome = "hg19")
}

}
