% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{findOverlaps,compartmap::CompartmentCall,compartmap::CompartmentCall-method}
\alias{findOverlaps,compartmap::CompartmentCall,compartmap::CompartmentCall-method}
\title{Find overlaps between \code{CompartmentCall} objects}
\usage{
## S7 method for classes <compartmap::CompartmentCall>, <compartmap::CompartmentCall>
findOverlaps(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)
}
\arguments{
\item{query}{A string vector of chromosomes to subset to}

\item{subject}{A \code{CompartmentCall} object}

\item{ maxgap,  minoverlap,  type, }{select, ignore.strand See \code{?findOverlaps} in the \code{GenomicRanges} package.}
}
\description{
Find overlaps between \code{CompartmentCall} objects
}
\concept{s7ranges}
