% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTransform.R
\name{fisherZ}
\alias{fisherZ}
\title{Fisher's Z transformation}
\usage{
fisherZ(cormat)
}
\arguments{
\item{cormat}{Pearson correlation matrix}
}
\value{
Fisher Z transformed Pearson correlations
}
\description{
\code{fisherZ} returns (squeezed) Fisher's Z transformed Pearson's r
}
\details{
This function returns (squeezed) Fisher's Z transformed Pearson's r
}
\examples{

#Generate a random binary (-1, 1) matrix
mat <- matrix(sample(c(1,-1), 10000, replace = TRUE), ncol = 100)

#Correct matrix diag
diag(mat) <- 1

#Transform
mat.transform <- fisherZ(mat)
}
