% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{fix_sign}
\alias{fix_sign}
\title{Correct the sign of the compartment call vector}
\usage{
fix_sign(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}

\item{na.rm}{Whether to remove NAs. The presence of NAs can reduce the
accuracy of the sign flip as the open vs closed gene densities cannot be
computed correctly}
}
\description{
Correct the sign of the compartment call vector
}
\concept{s7analysis}
