% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVD.R
\name{fixSign}
\alias{fixSign}
\title{Fix the sign of SVD singular vectors based on concordance between \strong{V}
direction and data direction}
\usage{
fixSign(mat, v)
}
\arguments{
\item{matrix}{A p x n input matrix}

\item{k}{Number of singular vectors to return}

\item{sing.vec}{Whether to return the right or left singular vector}
}
\value{
left or right singular vectors
}
\description{
Fix the sign of SVD singular vectors based on concordance between \strong{V}
direction and data direction
}
\details{
Modified from the reference below. Instead of doing \code{sign(dot_prod) * (dot_prod^2)} we just take the dot product of the 1st left singular vector
\strong{V} and the rows of the correlation matrix. Following the original
algorithm exactly results in the values always getting switched to positive
which is always what we want here. This also returns a sign 'dominance'
metric which the difference between the proportion of positive and negative
dot products: a value of 0.8 would mean that 90\% of the dot-products were in
the dominant direction: 0.9 - 0.1 = 0.8.

Resolving the sign ambiguity in the singular value decomposition, Journal of
Chemometrics 22(2):135-140
(https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/cem.1122)
Python reference: https://gist.github.com/David-Willo/1825bf9e8c30e13147e332734bcaebd5
Discussion of paper and why \code{abs()} is written in words but not in the algorithm:
https://www.mathworks.com/matlabcentral/fileexchange/22118-sign-correction-in-svd-and-pca
}
