% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flogit}
\alias{flogit}
\title{Helper function: squeezed logit}
\usage{
flogit(x, sqz = 0.000001)
}
\arguments{
\item{x}{a vector of values between 0 and 1 inclusive}

\item{sqz}{the amount by which to 'squeeze', default is .000001}
}
\value{
\preformatted{   a vector of values between -Inf and +Inf
}
}
\description{
Helper function: squeezed logit
}
