% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\alias{hg38.tx.gr}
\alias{hg19.tx.gr}
\alias{mm9.tx.gr}
\alias{mm10.tx.gr}
\title{Human and mouse genes as GRanges objects}
\format{
An object of class \code{GRanges} of length 35332.

An object of class \code{GRanges} of length 23056.

An object of class \code{GRanges} of length 21677.

An object of class \code{GRanges} of length 24528.
}
\usage{
data(hg38.tx.gr, package = "compartmap")

hg38.tx.gr

hg19.tx.gr

mm9.tx.gr

mm10.tx.gr
}
\description{
This object was generated using the
TxDb.Hsapiens.UCSC.hg19.knownGene/TxDb.Hsapiens.UCSC.hg38.knownGene/TxDb.Mmusculus.UCSC.mm9.knownGene/TxDb.Mmusculus.UCSC.mm10.knownGene
package. The script used for this object is found in the inst/scripts
directory
}
\author{
James Eapen \email{james.eapen@vai.org}
}
\keyword{datasets}
\keyword{reference_data}
