% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{get_assay}
\alias{get_assay}
\title{Get the assay of the \code{CompartmentCall}}
\usage{
get_assay(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get the assay of the \code{CompartmentCall}
}
\concept{s7getters}
