% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{get_bpnworkers}
\alias{get_bpnworkers}
\title{Get the total number of BiocParallelParam workers that will get used
From a BiocParallelParam or list of 2 BiocParallelParam objects}
\usage{
get_bpnworkers(bp)
}
\description{
Get the total number of BiocParallelParam workers that will get used
From a BiocParallelParam or list of 2 BiocParallelParam objects
}
\keyword{internal}
