% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{get_filter_threshold}
\alias{get_filter_threshold}
\title{Check the threshold at which an object's calls were filtered, returning 0 if not filtered}
\usage{
get_filter_threshold(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Check the threshold at which an object's calls were filtered, returning 0 if not filtered
}
