% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{get_md}
\alias{get_md}
\title{Compute Mahalanobis distances and pvalues - unfiltered}
\usage{
get_md(x, cov_method = c("base", "robust", "mcd"))
}
\arguments{
\item{x}{A \code{MultiCompartmentCall} object}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},
"robust": \code{robust::covRob()}, "mcd": \code{robust::covRob(estim = "mcd")}}

\item{alpha_level}{Significance level to use (default: 0.05)}
}
\description{
Compute Mahalanobis distances and pvalues - unfiltered
}
\concept{s7analysis}
