% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{get_name}
\alias{get_name}
\title{Get the name of the CompartmentCall object}
\usage{
get_name(x, ...)
}
\arguments{
\item{x}{A CompartmentCall object}
}
\description{
Get the name of the CompartmentCall object
}
