% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{get_nested_params}
\alias{get_nested_params}
\title{Set outer and inner params for nester parallelization
The outer param is across the input samples/columns and the second is for
bootstrapping. If \code{boot.parallel} is FALSE, the inner param is set to
\code{SerialParam}.}
\usage{
get_nested_params(BPPARAM, boot.parallel)
}
\description{
Set outer and inner params for nester parallelization
The outer param is across the input samples/columns and the second is for
bootstrapping. If \code{boot.parallel} is FALSE, the inner param is set to
\code{SerialParam}.
}
\keyword{internal}
