% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{get_resolution}
\alias{get_resolution}
\title{Get the resolution of the \code{CompartmentCall}}
\usage{
get_resolution(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get the resolution of the \code{CompartmentCall}
}
\concept{s7getters}
