% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArrayABsignal.R
\name{getArrayABsignal}
\alias{getArrayABsignal}
\title{Estimate A/B compartments from methylation array data}
\usage{
getArrayABsignal(
  obj,
  res = 1000000,
  chr = NULL,
  targets = NULL,
  preprocess = TRUE,
  parallel = TRUE,
  cores = 2,
  bootstrap = TRUE,
  num.bootstraps = 1000,
  boot.parallel = TRUE,
  boot.cores = 2,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  group = FALSE,
  other = NULL,
  array.type = c("hm450", "EPIC")
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{targets}{Samples/cells to shrink towards}

\item{preprocess}{Whether to preprocess the arrays prior to compartment inference}

\item{parallel}{Whether to run samples in parallel}

\item{cores}{How many cores to use when running samples in parallel}

\item{bootstrap}{Whether we should perform bootstrapping of inferred compartments}

\item{num.bootstraps}{How many bootstraps to run}

\item{boot.parallel}{Whether to run the bootstrapping in parallel}

\item{boot.cores}{How many cores to use for the bootstrapping}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{group}{Whether to treat this as a group set of samples}

\item{other}{Another arbitrary genome to compute compartments on}

\item{array.type}{What type of array is this ("hm450", "EPIC")}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{getArrayABsignal} returns estimated A/B compartments from methylation array data.
}
\examples{

if (requireNamespace("minfi", quietly = TRUE)) {
  data("array_data_chr14", package = "compartmap")
  array_compartments <- getArrayABsignal(
    array.data.chr14,
    chr="chr14",
    parallel=FALSE,
    bootstrap=FALSE,
    genome="hg19",
    array.type="hm450"
  )
}
}
