% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBinSparsity.R
\name{getBinSparsity}
\alias{getBinSparsity}
\title{Compute the sparsity of each bin that will be used to infer compartments}
\usage{
getBinSparsity(mat, row.ranges, chr = "chr14", res = 1e+05,
  min.depth = 1, scale.by.nonzero.bins = TRUE,
  filter.zero.bins = TRUE, scale.by.resolution = FALSE,
  species = c("human", "mouse"))
}
\arguments{
\item{mat}{Input matrix (samples as columns and loci as rows)}

\item{row.ranges}{The rowRanges or a GRanges object corresponding to genomic locations of loci}

\item{chr}{Which chromosome to operate on}

\item{res}{What resolution or bin size to use (default: 1e5)}

\item{min.depth}{Minimum depth to consider (default: 1)}

\item{scale.by.nonzero.bins}{Whether to scale the sparsity measure by the number of non-zero bins (default: TRUE)}

\item{filter.zero.bins}{Whether to filter non-zero bins (default: TRUE)}

\item{scale.by.resolution}{Whether to scale the per-locus sparsity by the resolution (default: FALSE)}

\item{species}{What species to use to determine chromosome sizes}
}
\value{
A list object containing either the scaled/summarized sparsity measures or a per-locus sparsity matrix and a corresponding GRanges object
}
\description{
Compute the sparsity of each bin that will be used to infer compartments
}
\examples{

}
