% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorMatrix.R
\name{getCorMatrix}
\alias{getCorMatrix}
\title{Calculate Pearson correlations of a binned matrix}
\usage{
getCorMatrix(binmat)
}
\arguments{
\item{binmat}{A binned matrix list object from getBinMatrix}
}
\value{
A list object to pass to getABSignal
}
\description{
This function is used to generate a list object to be passed to getABSignal
}
