% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMT.R
\name{getDenoisedMatrix}
\alias{getDenoisedMatrix}
\alias{getDenoisedCorMatrix}
\title{Wrapper to denoise a correlation matrix using a Random Matrix Theory approach}
\usage{
getDenoisedCorMatrix(
  obj,
  res = 1e+06,
  chr = "chr14",
  genome = c("hg19", "hg38", "mm9", "mm10"),
  iter = 2,
  targets = NULL,
  prior.means = NULL,
  assay = c("rna", "atac", "array")
)
}
\arguments{
\item{obj}{SummarizedExperiment object with rowRanges for each feature and colnames}

\item{res}{The resolution desired (default is a megabase 1e6)}

\item{chr}{Which chromosome to perform the denoising}

\item{genome}{Which genome (default is hg19)}

\item{iter}{How many iterations to perform denoising}

\item{targets}{Samples/cells to shrink towards}

\item{prior.means}{The means of the bin-level prior distribution (default will compute them for you)}

\item{assay}{What assay type this is ("rna", "atac")}
}
\value{
A denoised correlation matrix object for plotting with plotCorMatrix
}
\description{
Wrapper to denoise a correlation matrix using a Random Matrix Theory approach
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
denoised_cor_mat <- getDenoisedCorMatrix(k562_scrna_chr14, genome = "hg19", assay = "rna")
}
