% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomainInflections.R
\name{getDomainInflections}
\alias{getDomainInflections}
\title{A wrapper function to generate a GRanges object of chromatin domain inflection points}
\usage{
getDomainInflections(
  gr,
  what = "score",
  res = 1000000,
  chrs = c(paste0("chr", 1:22), "chrX"),
  genome = c("hg19", "hg38", "mm9", "mm10")
)
}
\arguments{
\item{gr}{Input GRanges object with mcols column corresponding to chromatin domains}

\item{what}{The name of the column containing the chromatin domain information}

\item{res}{What resolution the domains were called}

\item{chrs}{Which chromosomes to work on}

\item{genome}{Which genome does the input data come from}
}
\value{
A GRanges object of compartment inflection points
}
\description{
A wrapper function to generate a GRanges object of chromatin domain inflection points
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
chr14_domains <- scCompartments(k562_scrna_chr14,
  res = 1e6, genome = "hg19",
  group = TRUE, bootstrap = FALSE
)
chr14_domain_inflections <- getDomainInflections(chr14_domains, what = "cscore")
}
