% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenome}
\alias{getGenome}
\title{Get a GRanges object from bundled compartmap genomes}
\usage{
getGenome(genome = c("hg19", "hg38", "mm9", "mm10"), type = "genome")
}
\arguments{
\item{genome}{The desired genome to use ("hg19", "hg38", "mm9", "mm10")}

\item{type}{The type of data - full genome or open sea regions}
}
\value{
Granges of the genome
}
\description{
Get a GRanges object from bundled compartmap genomes
}
\examples{
hg19 <- getGenome(genome = "hg19")

}
\keyword{internal}
