% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlobalMeans.R
\name{getGlobalMeans}
\alias{getGlobalMeans}
\title{Get the global means of a matrix}
\usage{
getGlobalMeans(obj, targets = NULL, assay = c("array", "atac", "bisulfite"))
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{targets}{Column names or indices to indicate which samples to shrink towards}

\item{assay}{What type of assay the data are from}
}
\value{
A vector of global or targeted means
}
\description{
Get the global means of a matrix
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
array.global.means <- getGlobalMeans(array.data.chr14, assay = "array")

}
