% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatrixBlocks}
\alias{getMatrixBlocks}
\title{Get chunked sets of row-wise or column-wise indices of a matrix}
\usage{
getMatrixBlocks(mat, chunk.size = 100000, by.row = TRUE, by.col = FALSE)
}
\arguments{
\item{mat}{Input matrix}

\item{chunk.size}{The size of the chunks to use for coercion}

\item{by.row}{Whether to chunk in a row-wise fashion}

\item{by.col}{Whether to chunk in a column-wise fashion}
}
\value{
A set of chunked indices
}
\description{
Get chunked sets of row-wise or column-wise indices of a matrix
}
\examples{
#make a sparse binary matrix
library(Matrix)
m <- 100
n <- 1000
mat <- round(matrix(runif(m*n), m, n))
mat.sparse <- Matrix(mat, sparse = TRUE)

#get row-wise chunks of 10
chunks <- getMatrixBlocks(mat.sparse, chunk.size = 10)

}
