% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOpenSeas}
\alias{getOpenSeas}
\title{Gather open sea CpG from a GRanges of CpG islands}
\usage{
getOpenSeas(gr)
}
\arguments{
\item{gr}{Input GRanges of CpG islands}
}
\value{
GRanges object that can be used with filterOpenSea()
}
\description{
This function accepts a GRanges input of CpG islands that can
be derived from UCSC table browser and rtracklayer::import(yourbed.bed). It
resizes the inteval to create 4kb flanking open sea regions around the CpG
islands as defined by \href{https://doi.org/10.1186/s13059-015-0741-y}{Fortin and Hansen (Genome Biology, 2015)}.
}
\examples{
#cpgi <- rtracklayer::import(system.file("inst/extdata/mm10_cpgi.bed", package = "compartmap"))
#opensea_cpg <- getOpenSeas(cpgi)
}
