% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPMD}
\alias{getPMD}
\title{Retrieve probes, regions, etc. that fall within PMDs defined by Zhou et. al 2018
https://www.nature.com/articles/s41588-018-0073-4}
\usage{
getPMD(obj, genome = c("hg19", "hg38"))
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{genome}{Which genome annotations to use}
}
\value{
A subsetted SummarizedExperiment containing loci within PMDs
}
\description{
Retrieve probes, regions, etc. that fall within PMDs defined by Zhou et. al 2018
https://www.nature.com/articles/s41588-018-0073-4
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
pmd_loci <- getPMD(array.data.chr14, genome = "hg19")

}
