% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{Get the seqlengths of a chromosome from a given genome's GRanges}
\usage{
getSeqLengths(genome.gr, chr = "chr14")
}
\arguments{
\item{genome.gr}{A GRanges object of the genome (from \code{getGenome()})}

\item{chr}{What chromosome to extract the seqlengths of}
}
\value{
The seqlengths of a specific chromosome
}
\description{
The goal for this function is to eliminate the need to lug around
large packages when we only want seqlengths for things.
}
\examples{
hg19.chr14.seqlengths <- getSeqLengths(getGenome('hg19'), chr = "chr14")

}
