% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShrinkageTargets.R
\name{getShrinkageTargets}
\alias{getShrinkageTargets}
\title{Get the specified samples to shrink towards instead of the global mean}
\usage{
getShrinkageTargets(obj, group)
}
\arguments{
\item{obj}{Input matrix}

\item{group}{Sample names, colnames or column indices to use for targeted
shrinkage}
}
\value{
A matrix composed of samples to shrink towards
}
\description{
Get the specified samples to shrink towards instead of the global mean
}
\examples{
dummy <- matrix(rnorm(1000), ncol=25)
dummy.sub <- getShrinkageTargets(dummy, group = c(1,5,8,10))
}
