% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVD.R
\name{getSVD}
\alias{getSVD}
\title{Compute the SVD of a matrix using irlba}
\usage{
getSVD(matrix, k = 1, sing.vec = c("left", "right"))
}
\arguments{
\item{matrix}{A p x n input matrix}

\item{k}{Number of singular vectors to return}
}
\value{
A singular vector or matrix with sign corresponding to positive values
}
\description{
Compute the SVD of a matrix using irlba
}
\examples{

dummy <- matrix(rnorm(10000), ncol=25)
sing_vec <- getSVD(dummy, k = 1, sing.vec = "left")

}
