% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWGBSABsignal.R
\name{getWGBSABsignal}
\alias{getWGBSABsignal}
\title{Estimate A/B compartments from WGBS data}
\usage{
getWGBSABsignal(mat, res = 1e+06, globalMeanSet = NULL, noMean = FALSE,
  targets = NULL, parallel = FALSE, allchrs = FALSE, chr = NULL,
  regions = NULL, genome = "hg19", preprocess = TRUE, ...)
}
\arguments{
\item{mat}{A BSseq object}

\item{res}{Compartment resolution}

\item{globalMeanSet}{The global mean to shrink towards (calculated on the fly if NULL)}

\item{noMean}{Whether to shrink towards a global mean}

\item{targets}{Target samples to shrink towards}

\item{parallel}{Parallel estimation}

\item{allchrs}{Analyze all chromosomes (chr should be NULL if TRUE)}

\item{chr}{Chromosomes to analyze (list of arbitrary chromosomes)}

\item{regions}{GRanges object of loci locations (calculated on the fly)}

\item{genome}{Genome to analyze}

\item{preprocess}{Preprocessing with filtering, smoothing, and imputation}

\item{...}{Other args}
}
\value{
compartment estimates
}
\description{
\code{getWGBSABsignal} returns estimated A/B compartments from whole genome bisulfite sequencing data
}
\details{
This function estimates A/B compartments with some Dirichlet smoothing and shrinking towards a global mean of targets or across samples
}
\examples{


}
