% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{granges,compartmap::CompartmentCall-method}
\alias{granges,compartmap::CompartmentCall-method}
\title{Get GRanges of the \code{CompartmentCall}}
\usage{
## S7 method for class <compartmap::CompartmentCall>
granges(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get GRanges of the \code{CompartmentCall}
}
\concept{s7getters}
