% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTFIDF.R
\name{hdf5TFIDF}
\alias{hdf5TFIDF}
\title{Transform/normalize compartment calls using TF-IDF on HDF5-backed objects}
\usage{
hdf5TFIDF(h5, scale.factor = 100000, return.dense = FALSE, return.se = FALSE)
}
\arguments{
\item{h5}{SummarizedExperiment object, DelayedMatrix, or a normal matrix}

\item{scale.factor}{Scaling factor for the term-frequency (TF)}

\item{return.dense}{Whether to return a dense, in memory matrix}

\item{return.se}{Whether to return the TF-IDF matrix as a new assay in the SummarizedExperiment}
}
\value{
A TF-IDF transformed matrix of the same dimensions as the input
}
\description{
Transform/normalize compartment calls using TF-IDF on HDF5-backed objects
}
\examples{

m <- 1000
n <- 100
mat <- round(matrix(runif(m * n), m, n))
# Input needs to be a tall matrix
tfidf <- hdf5TFIDF(mat)

}
