% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.gr}
\alias{hg19.gr}
\title{hg19 seqlengths as a GRanges object}
\usage{
data(hg19.gr, package = "compartmap")
}
\description{
This object was generated using the Homo.sapiens package.
The script used for this object is found in the inst/scripts directory
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
