% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.gr}
\alias{hg38.gr}
\title{hg38 seqlengths as a GRanges object}
\usage{
data(hg38.gr, package = "compartmap")
}
\description{
This object was generated using the BSgenome.Hsapiens.UCSC.hg38 package.
The script used for this object is found in the inst/scripts directory
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
