% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.tx.gr}
\alias{hg38.tx.gr}
\title{hg38 genes as a GRanges object}
\usage{
data(hg38.tx.gr, package = "compartmap")
}
\description{
This object was generated using the TxDb.Hsapiens.UCSC.hg19.knownGene package.
The script used for this object is found in the inst/scripts directory
}
\author{
James Eapen \email{james.eapen@vai.org}
}
\keyword{data}
