% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importBigWig}
\alias{importBigWig}
\title{Import and optionally summarize a bigwig at a given resolution}
\usage{
importBigWig(
  bw,
  bins = NULL,
  summarize = FALSE,
  genome = c("hg19", "hg38", "mm9", "mm10")
)
}
\arguments{
\item{bw}{Path a bigwig file}

\item{bins}{Optional set of bins as a GRanges to summarize the bigwig to}

\item{summarize}{Whether to perform mean summarization}

\item{genome}{Which genome is the bigwig from ("hg19", "hg38", "mm9", "mm10")}
}
\value{
SummerizedExperiment object with rowRanges corresponding to summarized features
}
\description{
Import and optionally summarize a bigwig at a given resolution
}
