% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeKNN.R
\name{imputeKNN}
\alias{imputeKNN}
\title{Impute missing values/NAs with KNN}
\usage{
imputeKNN(obj, rowmax = 0.5, colmax = 0.8, k = 10, maxp = 1500,
  in.place = TRUE, drop.sparse.samps = TRUE, assay = c("array",
  "atac", "bisulfite"))
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{rowmax}{Maximum fraction of NAs that can exist in a row}

\item{colmax}{Maximum fraction of NAs that can exist in a column/sample}

\item{k}{Number of neighbors to be used in the imputation}

\item{maxp}{Largest block of regions/loci imputed using KNN}

\item{in.place}{Whether to modify the Beta/counts in place (default: TRUE)}

\item{drop.sparse.samps}{Whether to drop samples that are too sparse (default: TRUE)}

\item{assay}{The type of assay ("array", "atac", "bisulfite")}
}
\value{
Imputed data matrix that is added to the assays slot
}
\description{
Impute missing values/NAs with KNN
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
#impute
imputed <- imputeKNN(array.data.chr14, assay = "array")
}
