% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{is_filtered}
\alias{is_filtered}
\title{Check if the \code{CompartmentCall} was filtered.}
\usage{
is_filtered(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Check if the \code{CompartmentCall} was filtered.
}
\concept{s7getters}
