% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{is_unitarized}
\alias{is_unitarized}
\title{Check if the \code{CompartmentCall} was unitarized}
\usage{
is_unitarized(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Check if the \code{CompartmentCall} was unitarized
}
