% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k562_scatac_chr14}
\alias{k562_scatac_chr14}
\title{Example scATAC-seq data for compartmap}
\usage{
data(k562_scatac_chr14, package = "compartmap")
}
\description{
This data was generated using the data from the reference via bwa mem
and pre-processing the data using the csaw package.
}
\references{
\url{https://www.ncbi.nlm.nih.gov//geo/query/acc.cgi?acc=GSE99172}
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
