% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k562_scrna_chr14}
\alias{k562_scrna_chr14}
\title{Example scRNA-seq data for compartmap}
\usage{
data(k562_scrna_chr14, package = "compartmap")
}
\description{
This object was generated using the K562 data from the STORM-seq paper
and pre-processed using the scran and scater packages
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
