% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maskArrays}
\alias{maskArrays}
\title{Mask off probes using annotations MASK_general from Zhou et. al NAR 2017
https://doi.org/10.1093/nar/gkw967}
\usage{
maskArrays(obj, genome = c("hg19", "hg38"), array.type = NULL)
}
\arguments{
\item{obj}{Input SummarizedExperiment to be masked}

\item{genome}{Which genome annotations to use}

\item{array.type}{Array type (EPIC or hm450);
default is NULL and will try to figure it out on the fly}
}
\value{
A subsetted SummarizedExperiment without masked probes
}
\description{
Mask off probes using annotations MASK_general from Zhou et. al NAR 2017
https://doi.org/10.1093/nar/gkw967
}
\examples{

}
