% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{mat}
\alias{mat}
\title{Get the data matrix from a \code{MultiCompartmapCall} or \code{scCompartmapCall}}
\usage{
mat(x)
}
\arguments{
\item{x}{A \code{MultiCompartmapCall} object}
}
\description{
Get the data matrix from a \code{MultiCompartmapCall} or \code{scCompartmapCall}
}
