% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSmoother.R
\name{meanSmoother}
\alias{meanSmoother}
\title{Windowed mean smoother}
\usage{
meanSmoother(mat, k = 1, iters = 2, delta = 0, weights = NULL)
}
\arguments{
\item{mat}{Input data matrix: samples are columns, regions/loci are rows}

\item{k}{Number of windows to use (default = 1 to smooth with 1
window on either side of a position)}

\item{iters}{Number of iterations to smooth (default is 2)}

\item{delta}{Convergence threshhold (overrides iter if > 0; default is 0)}

\item{weights}{Weights, if using any (NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Smoothed data matrix
}\if{html}{\out{</div>}}
}
\description{
Windowed mean smoother
}
\examples{
dummy <- matrix(rnorm(10000), ncol = 25)
smooth.dummy <- meanSmoother(dummy)
smooth.dummy <- meanSmoother(dummy, iters = 3)
smooth.dummy <- meanSmoother(dummy, delta = 1e-3)

}
