% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.gr}
\alias{mm10.gr}
\title{mm10 seqlengths as a GRanges object}
\usage{
data(mm10.gr, package = "compartmap")
}
\description{
This object was generated using the Mus.musculus package.
The script used for this object is found in the inst/scripts directory
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
