% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.tx.gr}
\alias{mm10.tx.gr}
\title{mm10 genes as a GRanges object}
\usage{
data(mm10.tx.gr, package = "compartmap")
}
\description{
This object was generated using the TxDb.Mmusculus.UCSC.mm10.knownGene package.
The script used for this object is found in the inst/scripts directory
}
\author{
James Eapen \email{james.eapen@vai.org}
}
\keyword{data}
