% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm9.gr}
\alias{mm9.gr}
\title{mm9 seqlengths as a GRanges object}
\usage{
data(mm9.gr, package = "compartmap")
}
\description{
This object was generated using the BSgenome.Mmusculus.UCSC.mm9 package.
The script used for this object is found in the inst/scripts directory
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
