% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{MultiCompartmentCall}
\alias{MultiCompartmentCall}
\title{MultiCompartmentCall class}
\usage{
MultiCompartmentCall(ccalls, name, unitarized = FALSE, unitarize = FALSE)
}
\arguments{
\item{ccalls}{A list of CompartmapCalls to combine}

\item{name}{An identifier for this set of compartment calls}

\item{unitarized}{Whether the singular values have been unitarized}

\item{unitarize}{Whether to unitarize the singular values for each of the inputs calls}
}
\description{
An S7 class to hold multiple compartment calls at the same resolution with
their metadata and analysis methods. This can take a set of single
\code{CompartmentCall} or \code{CompartmapCall} objects.
}
