% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{names,compartmap::MultiCompartmapCall-method}
\alias{names,compartmap::MultiCompartmapCall-method}
\title{Get column names of \code{MultiCompartmapCall} matrix}
\usage{
## S7 method for class <compartmap::MultiCompartmapCall>
names(x)
}
\description{
Get column names of \code{MultiCompartmapCall} matrix
}
\concept{s7getters}
