% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openSeas.hg19}
\alias{openSeas.hg19}
\title{hg19 open sea CpG as a GRanges object}
\usage{
data(openSeas.hg19, package = "compartmap")
}
\description{
This object was generated using the Homo.sapiens package.
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
