% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openSeas.hg38}
\alias{openSeas.hg38}
\title{hg38 open sea CpG as a GRanges object}
\usage{
data(openSeas.hg38, package = "compartmap")
}
\description{
This object was generated using the BSgenome.Hsapiens.UCSC.hg38 package.
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
