% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openSeas.mm10}
\alias{openSeas.mm10}
\title{mm10 open sea CpG as a GRanges object}
\usage{
data(openSeas.mm10, package = "compartmap")
}
\description{
This object was generated using the Mus.musculus package.
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
