% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openSeas.mm9}
\alias{openSeas.mm9}
\title{mm9 open sea CpG as a GRanges object}
\usage{
data(openSeas.mm9, package = "compartmap")
}
\description{
This object was generated using the BSgenome.Mmusculus.UCSC.mm9 package.
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
