% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openSeas}
\alias{openSeas}
\alias{openSeas.hg38}
\alias{openSeas.hg19}
\alias{openSeas.mm9}
\alias{openSeas.mm10}
\title{Human and mouse open sea CpGs as GRanges objects}
\format{
An object of class \code{GRanges} of length 60378.

An object of class \code{GRanges} of length 33708.

An object of class \code{GRanges} of length 13225.

An object of class \code{GRanges} of length 13208.
}
\usage{
data(openSeas.hg38, package = "compartmap")

openSeas.hg38

openSeas.hg19

openSeas.mm9

openSeas.mm10
}
\description{
These objects were generated using the
BSgenome.Hsapiens.UCSC.hg19/BSgenome.Hsapiens.UCSC.hg38 and Mus.musculus
packages.
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{datasets}
\keyword{reference_data}
