% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{plot.compartmap::CompartmentCall}
\alias{plot.compartmap::CompartmentCall}
\title{Plot singular values of a \code{CompartmentCall} object}
\usage{
\method{plot}{`compartmap::CompartmentCall`}(
  x,
  ...,
  type = "line",
  label_coords = FALSE,
  res = "kb",
  width = 0.5,
  ylim = NULL
)
}
\arguments{
\item{x}{\code{CompartmentCall} or \code{CompartmapCall} object}

\item{...}{Placeholder for the \code{plot} generic - arguments have not effect}

\item{type}{Whether to plot the singular values as \code{"line"} or \code{"bar"} plots.}

\item{label_coords}{Label the x-axis with genomic coordinates. Uses a
numeric index when set to \code{FALSE}. Using coordinate labels can severely
crowd the x-axis, especially with Kb-resolution calls.}

\item{res}{The resolution to round the genomic coordinates to (kilobase:
"kb" or megabase: "mb")}

\item{width}{The width of the \code{geom_line} if \code{type = "line"} or the width
of the bar if \code{type = "bar"} in the plot}

\item{ylim}{Upper and lower bound for the y-axis}
}
\description{
Plot singular values of a \code{CompartmentCall} object
}
\concept{plotting}
