% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{plot.compartmap::MultiCompartmapCall}
\alias{plot.compartmap::MultiCompartmapCall}
\title{Plot singular values from a \code{MultiCompartmapCall} object}
\usage{
\method{plot}{`compartmap::MultiCompartmapCall`}(
  x,
  ...,
  type = "line",
  label_coords = FALSE,
  res = "mb",
  width = 0.5,
  ylim = NULL
)
}
\arguments{
\item{x}{\code{MultiCompartmapCall}}

\item{...}{Placeholder for the \code{plot} generic - arguments have not effect}

\item{type}{Whether to plot the singular values as \code{"line"}plots. Bar plots
will be facted by the individual \code{CompartmapCall} object names while the
line plots are overlayed.}

\item{label_coords}{Label the x-axis with genomic coordinates. Uses a
numeric index when set to \code{FALSE}. Using coordinate labels can severely
crowd the x-axis, especially with Kb-resolution calls.}

\item{res}{The resolution to round the genomic coordinates to (kilobase:
"kb" or megabase: "mb")}

\item{width}{The width of the \code{geom_line} if \code{type = "line"} or the width
of the bar if \code{type = "bar"} in the plot}

\item{ylim}{Upper and lower bound for the y-axis}
}
\description{
Plot singular values from a \code{MultiCompartmapCall} object
}
\concept{plotting}
