% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{plot.compartmap::MultiCompartmentCall}
\alias{plot.compartmap::MultiCompartmentCall}
\title{Plot singular values from a MultiCompartmentCall object}
\usage{
\method{plot}{`compartmap::MultiCompartmentCall`}(x)
}
\description{
Plot singular values from a MultiCompartmentCall object
}
