% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMT.R
\name{plot.RMT}
\alias{plot.RMT}
\title{Random matrix theory plot, stolen from \code{covmat}, which fell off of CRAN}
\usage{
\method{plot}{RMT}(x, y, ...)
}
\arguments{
\item{x}{model of type RMT obtained by fitting an RMT model to the data}

\item{y}{unused}

\item{...}{unused}
}
\description{
Eigenvalue plot
}
\details{
Plots eigenvalues of the correlation matrix and overlays the Marchenko-
Pastur density on top of it. There is a sharp cutoff for the density. We
are concerned with eigenvalues beyond this cutoff. Parameters used for
plotting are added to the plot.

TODO: replace the example with one from compartmap
TODO: consider making RMT an S4 class someday
}
\examples{
\dontrun{
 data("largereturn")
 model <- estRMT(largesymdata)
 plot(model)
}

}
\author{
Rohit Arora
}
