% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{plot_diff_compartments}
\alias{plot_diff_compartments}
\title{Compute differential compartments based on Mahalanobis distance}
\usage{
plot_diff_compartments(
  x,
  type = c("line", "bar"),
  cov_method = c("base", "robust", "mcd"),
  alpha_level = 0.05,
  show_md = TRUE,
  fill = "maroon",
  alpha = 0.5,
  ylim = c(-0.1, 0.1),
  label_ids = TRUE
)
}
\arguments{
\item{x}{A \code{MultiCompartmentCall} object}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},}

\item{alpha_level}{Significance level to use (default: 0.05)}

\item{show_md}{Whether to plot the Mahalanobis distance}

\item{fill}{Color of the \code{geom_rect}}

\item{alpha}{Transparency of the \code{geom_rect}}

\item{ylim}{The y-axis limits}

\item{label_ids}{Whether to differential bin ID labels}
}
\description{
Compute differential compartments based on Mahalanobis distance
}
\concept{s7analysis}
