% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArrayABsignal.R
\name{preprocessArrays}
\alias{preprocessArrays}
\title{Preprocess arrays for compartment inference}
\usage{
preprocessArrays(
  obj,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  other = NULL,
  array.type = c("hm450", "EPIC")
)
}
\arguments{
\item{obj}{Input SummarizedExperiment}

\item{genome}{What genome are we working on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}

\item{array.type}{What type of array is this ("hm450", "EPIC")}
}
\value{
A preprocessed SummarizedExperiment to compute compartments
}
\description{
Preprocess arrays for compartment inference
}
\examples{
if (require(minfiData)) {
  grSet <- mapToGenome(ratioConvert(preprocessNoob(RGsetEx.sub)))
  preprocessArrays(grSet)
} 
}
