% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSABsignal.R
\name{preprocessBSseq}
\alias{preprocessBSseq}
\title{Preprocess BS-seq for compartment inference}
\usage{
preprocessBSseq(obj, genome = c("hg19", "hg38", "mm9", "mm10"), other = NULL)
}
\arguments{
\item{obj}{Input SummarizedExperiment}

\item{genome}{What genome are we working on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}
}
\value{
A preprocessed SummarizedExperiment to compute compartments
}
\description{
Preprocess BS-seq for compartment inference
}
\examples{

}
