% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWGBSABsignal.R
\name{preprocessWGBS}
\alias{preprocessWGBS}
\alias{preprocessArrays}
\title{Preprocess BS-seq for compartment inference}
\usage{
preprocessArrays(obj, res = 1e+06, genome = c("hg19", "hg38", "mm9",
  "mm10"), other = NULL)
}
\arguments{
\item{obj}{Input SummarizedExperiment}

\item{res}{Compartment resolution in bp}

\item{genome}{What genome are we working on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}
}
\value{
A preprocessed SummarizedExperiment to compute compartments
}
\description{
Preprocess BS-seq for compartment inference
}
\examples{

}
