% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{seqinfo,compartmap::CompartmentCall-method}
\alias{seqinfo,compartmap::CompartmentCall-method}
\title{Get Seqinfo of the \code{CompartmentCall}}
\usage{
## S7 method for class <compartmap::CompartmentCall>
seqinfo(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get Seqinfo of the \code{CompartmentCall}
}
\concept{s7getters}
