% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqlengths}
\alias{seqlengths}
\alias{hg38.gr}
\alias{hg19.gr}
\alias{mm9.gr}
\alias{mm10.gr}
\title{Human and mouse seqlengths as GRanges objects}
\format{
An object of class \code{GRanges} of length 455.

An object of class \code{GRanges} of length 93.

An object of class \code{GRanges} of length 35.

An object of class \code{GRanges} of length 66.
}
\usage{
data(hg38.gr, package = "compartmap")

hg38.gr

hg19.gr

mm9.gr

mm10.gr
}
\description{
These objects were generated using the
BSgenome.Hsapiens.UCSC.hg19/BSgenome.Hsapiens.UCSC.hg38 and Mus.musculus
packages. The script used is found in the inst/scripts directory
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{datasets}
\keyword{reference_data}
