% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{seqlevels,compartmap::CompartmentCall-method}
\alias{seqlevels,compartmap::CompartmentCall-method}
\title{Get 'seqlevels' of the \code{CompartmentCall}}
\usage{
## S7 method for class <compartmap::CompartmentCall>
seqlevels(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get 'seqlevels' of the \code{CompartmentCall}
}
\concept{s7getters}
