% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkBins.R
\name{shrinkBins}
\alias{shrinkBins}
\title{Employ an eBayes shrinkage approach for bin-level estimates for A/B inference}
\usage{
shrinkBins(x, prior.means = NULL, chr = NULL, res = 1e+06,
  targets = NULL, resample = FALSE, assay = c("array", "atac",
  "bisulfite"))
}
\arguments{
\item{x}{Input SummarizedExperiment object}

\item{prior.means}{The means of the bin-level prior distribution}

\item{chr}{The chromosome to operate on}

\item{res}{Resolution to perform the binning}

\item{targets}{The column/sample/cell names to shrink towards}

\item{resample}{Whether this is for resampling purposes}

\item{assay}{What assay type this is ("array", "atac", "bisulfite")}
}
\value{
A list object to pass to getCorMatrix
}
\description{
\code{shrinkBins} returns shrunken bin-level estimates
}
\details{
This function computes shrunken bin-level estimates using a James-Stein estimator, reformulated as an eBayes procedure
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
#impute to remove NAs
imputed.array <- imputeKNN(array.data.chr14, assay = "array")
#get the shrunken binned M-values
shrunken.bin.array <- shrinkBins(imputed.array, chr = "chr14", assay = "array")

}
