% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkBins.R
\name{shrinkBinsArray}
\alias{shrinkBinsArray}
\title{Employ an eBayes shrinkage approach for bin-level estimates for A/B inference in methylation arrays}
\usage{
shrinkBinsArray(x, prior.means = NULL, chr = NULL, res = 1e+06, ...)
}
\arguments{
\item{x}{a RangedSummarizedExperiment}

\item{prior.means}{the means of the bin-level prior distribution}

\item{chr}{the chromosome to operate on}

\item{res}{resolution to perform the binning}

\item{...}{other parameters to pass to the internal functions}
}
\value{
A list object to pass to getCorMatrix
}
\description{
\code{shrinkBinsArray} returns shrunken bin-level estimates from methylation arrays
}
\details{
This function computes shrunken bin-level estimates using a James-Stein estimator, reformulated as an eBayes procedure
}
\examples{

}
