% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{SingleCellCompartmentCall}
\alias{SingleCellCompartmentCall}
\title{SingleCellCompartmentCall class}
\usage{
SingleCellCompartmentCall(
  ccalls,
  res,
  name,
  unitarized = FALSE,
  unitarize = FALSE
)
}
\arguments{
\item{ccalls}{A RageedExperiment of single-cell compartment calls}

\item{name}{An identifier for this set of compartment calls}

\item{unitarized}{Whether the singular values have been unitarized}

\item{unitarize}{Whether to unitarize the singular values for each of the inputs calls}
}
\description{
An S7 class to hold multiple single-cell level compartment calls at the same
resolution with their metadata and analysis methods. This takes a
RaggedExperiment from \code{scCompartments()} as its input.
}
