% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparseToDenseMatrix}
\alias{sparseToDenseMatrix}
\title{Convert a sparse matrix to a dense matrix in a block-wise fashion}
\usage{
sparseToDenseMatrix(
  mat,
  blockwise = TRUE,
  by.row = TRUE,
  by.col = FALSE,
  chunk.size = 1e+05,
  parallel = FALSE,
  cores = 2
)
}
\arguments{
\item{mat}{Input sparse matrix}

\item{blockwise}{Whether to do the coercion in a block-wise manner}

\item{by.row}{Whether to chunk in a row-wise fashion}

\item{by.col}{Whether to chunk in a column-wise fashion}

\item{chunk.size}{The size of the chunks to use for coercion}

\item{parallel}{Whether to perform the coercion in parallel}

\item{cores}{The number of cores to use in the parallel coercion}
}
\value{
A dense matrix of the same dimensions as the input
}
\description{
Convert a sparse matrix to a dense matrix in a block-wise fashion
}
\examples{
#make a sparse binary matrix
library(Matrix)
m <- 100
n <- 1000
mat <- round(matrix(runif(m*n), m, n))
mat.sparse <- Matrix(mat, sparse = TRUE)

#coerce back
mat.dense <- sparseToDenseMatrix(mat.sparse, chunk.size = 10)

#make sure they are the same dimensions
dim(mat) == dim(mat.dense)

#make sure they are the same numerically
all(mat == mat.dense)

}
