% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss3_umi_sce}
\alias{ss3_umi_sce}
\title{Example SMART-seq3 scRNA-seq data for compartmap}
\usage{
data(ss3_umi_sce, package = "compartmap")
}
\description{
Only keep chromosome 22 for the example
}
\details{
This object was generated using the HEK293T data from the SMART-seq3 paper
}
\references{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-8735/}
}
\author{
Benjamin K Johnson \email{ben.johnson@vai.org}
}
\keyword{data}
