% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{[.compartmap::CompartmentCall}
\alias{[.compartmap::CompartmentCall}
\title{Subset rows of a \code{CompartmentCall} object}
\usage{
\method{[}{`compartmap::CompartmentCall`}(x, i = NULL)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}

\item{i}{Row indices to subset}
}
\description{
Subset rows of a \code{CompartmentCall} object
}
\keyword{internal}
