% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{[.compartmap::MultiCompartmapCall}
\alias{[.compartmap::MultiCompartmapCall}
\title{Subset \code{MultiCompartmapCall} object}
\usage{
\method{[}{`compartmap::MultiCompartmapCall`}(x, i = NULL, j = NULL)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}

\item{i}{Rows indices to subset}

\item{j}{Column indices or names to subset}
}
\description{
Subset \code{MultiCompartmapCall} object
}
\keyword{internal}
