% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{subset_chr}
\alias{subset_chr}
\title{Subset the CompartmentCall object by chromosome}
\usage{
subset_chr(x, chr)
}
\arguments{
\item{x}{A CompartmentCall object}

\item{chr}{A string vector of chromosomes to subset to}
}
\description{
Subset the CompartmentCall object by chromosome
}
