% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTFIDF.R
\name{transformTFIDF}
\alias{transformTFIDF}
\title{Transform/normalize counts using TF-IDF}
\usage{
transformTFIDF(mat, scale.factor = 100000, count.min = 0, count.max = 1)
}
\arguments{
\item{mat}{n x p input matrix (n = samples/cells; p = rna counts)}

\item{scale.factor}{Scaling factor for the term-frequency (TF)}

\item{count.min}{The minimum expression count used for TF-IDF. Binarizes
when \code{count.min} = 0 and \code{count.max} = 1.}

\item{count.max}{The maximum expression count used for TF-IDF. Binarizes
when \code{count.min} = 0 and \code{count.max} = 1.
binarizes the matrix. A \code{cap} value greater than 1 will cap counts at that
value.}
}
\value{
A TF-IDF transformed matrix of the same dimensions as the input
}
\description{
Transform/normalize counts using TF-IDF
}
\details{
This function and its helpers were modeled after or taken from:
\itemize{
\item http://andrewjohnhill.com/images/posts/2019-5-6-dimensionality-reduction-for-scatac-data/analysis.html
\item https://divingintogeneticsandgenomics.rbind.io/post/clustering-scatacseq-data-the-tf-idf-way/
}
}
\examples{
m <- 1000
n <- 100
mat <- round(matrix(runif(m * n), m, n))
# Input needs to be a tall matrix
tfidf <- transformTFIDF(mat)

}
