% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{unitarize}
\alias{unitarize}
\title{Unitarize the singular values in a \code{CompartmentCall} or \code{MultiCompartmentCall}}
\usage{
unitarize(x, medianCenter = TRUE)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}

\item{medianCenter}{Whether to center the singular values on their median}
}
\description{
Unitarize the singular values in a \code{CompartmentCall} or \code{MultiCompartmentCall}
}
