% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{verify_bp}
\alias{verify_bp}
\title{Verify that the input BiocParallelParam is valid}
\usage{
verify_bp(bp)
}
\arguments{
\item{A}{BiocParallelParam or list of 2 BiocParallelParam objects}
}
\value{
TRUE if the total \code{bpnworkers} in the input does not exceed
available resources as defined by \code{parallelly::availableCores()}
}
\description{
Verify that the input BiocParallelParam is valid
}
\keyword{internal}
