% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{verify_workers}
\alias{verify_workers}
\title{Verify that requested thread count is not higher than available}
\usage{
verify_workers(n_workers)
}
\arguments{
\item{thread_count}{The number of workers to check availability}
}
\value{
TRUE if the requested \code{thread_count} does not exceed available
resources as defined by \code{parallelly::availableCores()}
}
\description{
Verify that requested thread count is not higher than available
}
\keyword{internal}
