% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifyAssayNames}
\alias{verifyAssayNames}
\title{Check that the input SummarizedExperiment object has the right assays}
\usage{
verifyAssayNames(se, assay)
}
\arguments{
\item{se}{Input SummarizedExperiment object}

\item{assay}{The assay type}
}
\value{
Error if the right assay type is not present, NULL if it is
}
\description{
Check that the input SummarizedExperiment object has the right assays
}
\keyword{internal}
