% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verifyCoords}
\alias{verifyCoords}
\title{Throw error if assay does not contain coordinates}
\usage{
verifyCoords(obj)
}
\arguments{
\item{obj}{Input object}
}
\description{
Throw error if assay does not contain coordinates
}
\keyword{internal}
