% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verifySE}
\alias{verifySE}
\title{Check if the assay is a SummarizedExperiment}
\usage{
verifySE(obj)
}
\arguments{
\item{obj}{Input object}
}
\description{
Check if the assay is a SummarizedExperiment
}
\keyword{internal}
