% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{agreement}
\alias{agreement}
\title{Compute agreement between compartment calls}
\usage{
agreement(x)
}
\arguments{
\item{x}{A \code{MultiCompartmapCall} object}
}
\description{
Calculates the proportion of calls with the same sign for every pair of
calls in a \code{MultiCompartmapCall} object.
}
