% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{agreement}
\alias{agreement}
\title{Compute sign agreement of a matrix of singular values}
\usage{
agreement(mat)
}
\arguments{
\item{mat}{A matrix of compartment calls where each column is a single-cell
or grouped set of singular values}
}
\description{
Compute sign agreement of a matrix of singular values
}
