% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrestiCoullCI.R
\name{agrestiCoullCI}
\alias{agrestiCoullCI}
\title{Agresti-Coull confidence interval for a binomial proportion}
\usage{
agrestiCoullCI(n1, n0, q)
}
\arguments{
\item{n1}{number of successes/ones}

\item{n0}{number of failures/zeroes}

\item{q}{quantile for eventual CI (e.g. 0.95 for a 95\\% binomial CI)}
}
\value{
the \"approximate\" (q*100)\\% confidence interval for (p|n1,n0,q)
}
\description{
Agresti-Coull confidence interval for a binomial proportion
}
\examples{
binom.ci <- agrestiCoullCI(10, 3, 0.95)

}
