% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayCompartments.R
\name{arrayCompartments}
\alias{arrayCompartments}
\title{Estimate A/B compartments from methylation array data}
\usage{
arrayCompartments(
  obj,
  res = 1000000,
  chr = NULL,
  group = FALSE,
  targets = NULL,
  bootstrap = TRUE,
  num.bootstraps = 1000,
  preprocess = TRUE,
  array.type = c("hm450", "EPIC"),
  genome = c("hg19", "hg38", "mm9", "mm10"),
  other = NULL,
  boot.parallel = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{group}{Whether to treat this as a group set of samples}

\item{targets}{Samples/cells to shrink towards}

\item{bootstrap}{Whether we should perform bootstrapping of inferred compartments}

\item{num.bootstraps}{How many bootstraps to run}

\item{preprocess}{Whether to preprocess the arrays prior to compartment inference}

\item{array.type}{What type of array is this ("hm450", "EPIC")}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}

\item{boot.parallel}{Whether to run the bootstrapping in parallel. See details.}

\item{BPPARAM}{BiocParallelParam object to use for parallelization. See details.}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{arrayCompartments} returns estimated A/B compartments from methylation array data.
}
\examples{

if (requireNamespace("minfi", quietly = TRUE)) {
  data("array_data_chr14", package = "compartmap")
  array_compartments <- arrayCompartments(
    array.data.chr14,
    chr="chr14",
    parallel=FALSE,
    bootstrap=FALSE,
    genome="hg19",
    array.type="hm450"
  )
}
}
