% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanAssayRows}
\alias{cleanAssayRows}
\title{Remove rows with NAs exceeding a threshold}
\usage{
cleanAssayRows(se, rowmax = 0.5, assay = c("array", "bisulfite"))
}
\arguments{
\item{se}{Input SummarizedExperiment object}

\item{rowmax}{The maximum NAs allowed in a row as a fraction}

\item{assay}{The type of assay we are working with}
}
\value{
A filtered matrix
}
\description{
Remove rows with NAs exceeding a threshold
}
\examples{
if (requireNamespace("minfi", quietly = TRUE)) {
  data("array_data_chr14", package = "compartmap")
  cleanAssayRows(array.data.chr14, assay = "array")
}
}
