% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{CompartmentCall}
\alias{CompartmentCall}
\title{CompartmentCall class (experimental)}
\usage{
CompartmentCall(pc, res, gr, name = NULL, unitarized = FALSE)
}
\arguments{
\item{pc}{The singular values from a compartment call}

\item{res}{The binning resolution used}

\item{gr}{The GRanges of the bins}

\item{name}{An identifier for this copmartment call. Important to set if you
are making a MultiCompartmentCall}

\item{unitarized}{Whether the singular values have been unitarized}
}
\description{
An S7 class to hold a compartment call with its metadata and analysis
methods. This is primarily a parent class from which the more used
\code{CompartmapCall} and \code{MultiCompartmentCall} are derived.
}
