% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{compute_mahalanobis}
\alias{compute_mahalanobis}
\title{Get the Mahalanobis distances for the compartment scores}
\usage{
compute_mahalanobis(
  mat,
  cov_method = c("base", "robust", "mcd"),
  alpha_level = 0.05,
  fdr = TRUE
)
}
\arguments{
\item{mat}{Matrix of singular values}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},
"robust": \code{robust::covRob()}, "mcd": \code{robust::covRob(estim = "mcd")}}

\item{alpha_level}{Significance level to use (default: 0.05)}

\item{fdr}{Whether to perform Benjamini-Hochberg false discovery correction}
}
\description{
Get the Mahalanobis distances for the compartment scores
}
\details{
The resulting data frame contains the mahalanobis distances and p-values
computed based on the squared mahalanobis distances on the Chi-squared
distribution.
}
\concept{diffCompartments}
