% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{dc_plotter}
\alias{dc_plotter}
\title{Show differential compartments as overlaid \code{ggplot2::geom_rect()} over significant distances}
\usage{
dc_plotter(
  ccall_pd,
  md,
  select,
  type = c("line", "bar"),
  show_md = TRUE,
  alpha_level = 0.05,
  shade = "maroon",
  alpha = 0.5,
  ylim = c(-0.5, 0.5),
  xlim = NULL,
  label_ids = TRUE,
  color = NULL,
  fill = NULL,
  linewidth = 0.5
)
}
\arguments{
\item{ccall_pd}{The \verb{@df} slot of a \code{MultiCompartmapCall} object}

\item{md}{data.table output from \code{compute_mahalanobis()}}

\item{select}{Column name or index to compute differential compartments on
Set this to plot the compartment scores of all input columns but show
differential compartments based only on the provided columns.}

\item{show_md}{Whether to plot the Mahalanobis distance}

\item{alpha_level}{Significance level to use (default: 0.05)}

\item{shade}{Color of the \code{geom_rect} used to shade differential bins}

\item{alpha}{Transparency of the \code{geom_rect}}

\item{ylim}{The y-axis limits}

\item{xlim}{The x-axis limits}

\item{label_ids}{Whether to differential bin ID labels}

\item{color}{The line colors used for the each sample}

\item{fill}{The colors used for positive and negative values in bar plots}

\item{linewidth}{The width of lines in the line plot}
}
\description{
Show differential compartments as overlaid \code{ggplot2::geom_rect()} over significant distances
}
\examples{
v <- c(1:4, 5, 7:9)
get_sequential_idx(v)
}
\concept{diffCompartments}
\keyword{internal}
