% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMT.R
\name{denoiseCorMatrix}
\alias{denoiseCorMatrix}
\title{Denoising of Covariance matrix using Random Matrix Theory}
\usage{
denoiseCorMatrix(cormat, rows, cols)
}
\arguments{
\item{cormat}{Correlation matrix of the data matrix}

\item{rows}{Row count of the data matrix}

\item{cols}{Column count of the data matrix}
}
\description{
Denoising of Covariance matrix using Random Matrix Theory
}
\details{
Given a random matrix \eqn{X_{m\times n}} with independently and identically
distributed values, the eigendecomposition of its covariance matrix \eqn{C}
is

\deqn{C = VDV^T}

where \eqn{D} is a diagonal matrix of its eigenvalues and $V$ contains
corresponding eigenvectors of \eqn{C}. The Marchenko-Pastur distribution
describes the distribution of random matrix eigenvalues, specifying that
they fall with the bounds

\deqn{lambda_{\pm }=\sigma ^{2}\left(1\pm {\sqrt{\frac{m}{n}}}\right)^{2}}

where \eqn{\sigma^2} is the variance, equal to 1 in the case of correlation
matrices and \eqn{m/n} the aspect ratio of the matrix. As \eqn{\lambda_+} is
the maximum expected eigenvalue in the Marchenko-Pastur distribution and
eigenvalues greater than \eqn{\lambda_+} are unlikely to derive from noise,
we interpret these as signal. Given \eqn{k} eigenvalues greater than
\eqn{\lambda_+}, we can reconstruct a `denoised' correlation matrix

\deqn{C' = V_kD_kV_k^T}

where \eqn{V_k} and \eqn{D_k} denote the eigenvectors and eigenvalues
respectively corresponding to the \eqn{k} largest eigenvalues. In effect, we
start by assuming that the correlation matrix is random noise to find the
upper bound of its Marchenko-Pastur distribution. Eigenvalues above this
bound violate this assumption and allow us to reconstruct the correlation
matrix excluding the elements representing random noise.
}
