% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{DF}
\alias{DF}
\title{Get the \verb{@dt} slot from a CompartmentCall object}
\usage{
DF(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get the \verb{@dt} slot from a CompartmentCall object
}
