% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R, R/MultiCompartmapCall.R
\name{dim,compartmap::CompartmentCall-method}
\alias{dim,compartmap::CompartmentCall-method}
\alias{dim,compartmap::MultiCompartmapCall-method}
\title{Get dimensions of \code{CompartmapCall} matrix}
\usage{
## S7 method for class <compartmap::CompartmentCall>
dim(x)

## S7 method for class <compartmap::MultiCompartmapCall>
dim(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Get dimensions of \code{CompartmapCall} matrix

Get dimensions of \code{MultiCompartmapCall} matrix
}
\concept{s7getters}
