% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{filter}
\alias{filter}
\title{Filter to bins with call values greater than or equal to a threshold value}
\usage{
filter(x, threshold = 0.02)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}

\item{threshold}{The absolute value to use for filtering. Rows where any
value is less than this threshold are dropped}
}
\description{
Filter to bins with call values greater than or equal to a threshold value
}
