% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterOpenSea}
\alias{filterOpenSea}
\alias{getOpenSeas}
\title{Filter to open sea CpG loci}
\usage{
filterOpenSea(obj, genome = c("hg19", "hg38", "mm10", "mm9"), other = NULL)

getOpenSeas(gr)
}
\arguments{
\item{obj}{Input SummarizedExperiment or GRanges object}

\item{genome}{Which genome to filter}

\item{other}{GRanges of open sea regions (TODO)}

\item{gr}{Input GRanges of CpG islands}
}
\value{
Filtered to open sea CpG loci

GRanges object that can be used with filterOpenSea()
}
\description{
This function accepts a GRanges input of CpG islands that can
be derived from UCSC table browser and rtracklayer::import(yourbed.bed)
}
\examples{
if (require(minfi)) {
  data("meth_array_450k_chr14", package = "compartmap")
  opensea <- filterOpenSea(array.data.chr14, genome = "hg19")
}
#cpgi <- rtracklayer::import(system.file("inst/extdata/mm10_cpgi.bed", package = "compartmap"))
#opensea_cpg <- getOpenSeas(cpgi)
}
