% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{fix_sign}
\alias{fix_sign}
\title{Correct the sign of the compartment call vector}
\usage{
fix_sign(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\description{
Correct the sign of the compartment call vector
}
\concept{s7analysis}
