% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCompartments.R
\name{fixCompartments}
\alias{fixCompartments}
\title{Invert, or "fix", compartments that have a minimum confidence score (1-min.conf)}
\usage{
fixCompartments(obj, min.conf = 0.8, parallel = FALSE, cores = 1)
}
\arguments{
\item{obj}{Input RaggedExperiment or output of condenseSE}

\item{min.conf}{Minimum confidence score to use}

\item{parallel}{Whether to run in parallel}

\item{cores}{How many cores to use if running in parallel}
}
\value{
A "fixed" set of compartments
}
\description{
Invert, or "fix", compartments that have a minimum confidence score (1-min.conf)
}
\examples{
 
}
