% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{flip}
\alias{flip}
\title{Flip the singular values signs in a CompartmentCall}
\usage{
flip(x, ...)
}
\arguments{
\item{x}{A CompartmentCall object}
}
\description{
Flip the singular values signs in a CompartmentCall
}
