% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCompartments.R
\name{flipper}
\alias{flipper}
\title{Helper to invert, or "fix", compartments that have a minimum confidence score (1-min.conf)}
\usage{
flipper(input_obj, min.conf)
}
\arguments{
\item{min.conf}{Minimum confidence score to use}

\item{obj}{Input RaggedExperiment or output of condenseSE}
}
\value{
A "fixed" set of compartments
}
\description{
Helper to invert, or "fix", compartments that have a minimum confidence score (1-min.conf)
}
\keyword{internal}
