% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{get_dc}
\alias{get_dc}
\title{Compute differential compartments based on Mahalanobis distance}
\usage{
get_dc(x, cov_method = c("base", "robust", "mcd"), alpha_level = 0.05)
}
\arguments{
\item{x}{A \code{MultiCompartmentCall} object}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},
"robust": \code{robust::covRob()}, "mcd": \code{robust::covRob(estim = "mcd")}}

\item{alpha_level}{Significance level to use (default: 0.05)}
}
\description{
Compute differential compartments based on Mahalanobis distance
}
\concept{diffCompartments}
