% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{get_sequential_idx}
\alias{get_sequential_idx}
\title{Get the start and end of sequences within a vector of integers}
\usage{
get_sequential_idx(v)
}
\arguments{
\item{v}{Indices of significant Mahalanobis distances}
}
\description{
It returns a data.table where rows may be duplicated to allow for additional
row-specific data if necessary
}
\examples{
v <- c(1:4, 5, 7:9)
get_sequential_idx(v)
}
\concept{diffCompartments}
