% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getABSignal.R
\name{getABSignal}
\alias{getABSignal}
\title{Calculate Pearson correlations of smoothed eigenvectors}
\usage{
getABSignal(x, k = 2, iter = 2)
}
\arguments{
\item{x}{A list object from getCorMatrix}

\item{k}{Value of k for smoothing (default = 2)}

\item{iter}{Number of iterations for moving average smoothing (default = 2)}
}
\value{
A list x to pass to getABSignal
}
\description{
This function is used to generate a list x to be passed to getABSignal
}
