% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArrayABsignal.R
\name{getArrayABsignal}
\alias{getArrayABsignal}
\title{Estimate A/B compartments from methylation array data}
\usage{
getArrayABsignal(obj, res = 1e+06, parallel = TRUE, chr = NULL,
  targets = NULL, preprocess = TRUE, cores = 1, genome = c("hg19",
  "hg38", "mm9", "mm10"), other = NULL, array.type = c("hm450",
  "EPIC"), bulk = FALSE, boot.parallel = FALSE, boot.cores = 1)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{parallel}{Whether to run samples in parallel}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{targets}{Samples/cells to shrink towards}

\item{preprocess}{Whether to preprocess the arrays prior to compartment inference}

\item{cores}{How many cores to use when running samples in parallel}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}

\item{array.type}{What type of array is this ("hm450", "EPIC")}

\item{bulk}{Whether to treat this as a bulk set of samples}

\item{boot.parallel}{Whether to run the bootstrapping in parallel}

\item{boot.cores}{How many cores to use for the bootstrapping}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{getArrayABsignal} returns estimated A/B compartments from methylation array data.
}
\examples{

}
