% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArrayABsignal.R
\name{getArrayABsignal}
\alias{getArrayABsignal}
\title{Estimate A/B compartments from methylation array data}
\usage{
getArrayABsignal(obj, res = 1e+06, parallel = FALSE, allchrs = FALSE,
  chr = NULL, targets = NULL, ...)
}
\arguments{
\item{obj}{Input GenomicRatioSet object}

\item{res}{Compartment resolution (in bp)}

\item{parallel}{Should the inference be done in parallel?}

\item{allchrs}{Whether all autosomes should be used for A/B inference}

\item{chr}{Specific chromosomes to analyze}

\item{targets}{Specify samples to use as shrinkage targets}

\item{...}{Additional arguments}
}
\value{
A p x n matrix (samples as columns and compartments as rows) of compartments
}
\description{
\code{getArrayABsignal} returns estimated A/B compartments from methylation array data.
}
\details{
This function is modified from the minfi::compartments to infer A/B compartments from array data
}
