% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArrayABsignal.R
\name{getArrayABsignal}
\alias{getArrayABsignal}
\title{Estimate A/B compartments from methylation array data}
\usage{
getArrayABsignal(
  obj,
  res = 1e+06,
  parallel = TRUE,
  chr = NULL,
  targets = NULL,
  preprocess = TRUE,
  cores = 2,
  bootstrap = TRUE,
  num.bootstraps = 1000,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  other = NULL,
  array.type = c("hm450", "EPIC"),
  group = FALSE,
  boot.parallel = TRUE,
  boot.cores = 2
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{parallel}{Whether to run samples in parallel}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{targets}{Samples/cells to shrink towards}

\item{preprocess}{Whether to preprocess the arrays prior to compartment inference}

\item{cores}{How many cores to use when running samples in parallel}

\item{bootstrap}{Whether we should perform bootstrapping of inferred compartments}

\item{num.bootstraps}{How many bootstraps to run}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{other}{Another arbitrary genome to compute compartments on}

\item{array.type}{What type of array is this ("hm450", "EPIC")}

\item{group}{Whether to treat this as a group set of samples}

\item{boot.parallel}{Whether to run the bootstrapping in parallel}

\item{boot.cores}{How many cores to use for the bootstrapping}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{getArrayABsignal} returns estimated A/B compartments from methylation array data.
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
array_compartments <- getArrayABsignal(array.data.chr14, parallel=F, chr="chr14", bootstrap=F, genome="hg19", array.type="hm450")
}
