% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAssayNames}
\alias{getAssayNames}
\title{Get the assay names from a SummarizedExperiment object}
\usage{
getAssayNames(se)
}
\arguments{
\item{se}{Input SummarizedExperiment object}
}
\value{
The names of the assays
}
\description{
Get the assay names from a SummarizedExperiment object
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
getAssayNames(k562_scrna_chr14)
}
