% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATACABsignal.R
\name{getATACABsignal}
\alias{getATACABsignal}
\title{Estimate A/B compartments from ATAC-seq data}
\usage{
getATACABsignal(obj, res = 1e+06, parallel = FALSE, allchrs = FALSE, ...)
}
\arguments{
\item{obj}{Input GenomicRatioSet object}

\item{res}{Compartment resolution (in bp)}

\item{parallel}{Should the inference be done in parallel?}

\item{allchrs}{Whether all autosomes should be used for A/B inference}

\item{...}{Additional arguments}
}
\value{
A p x n matrix (samples as columns and compartments as rows) of compartments
}
\description{
\code{getATACABsignal} returns estimated A/B compartments from methylation array data.
}
\details{
This function estimates A/B compartments shrinking towards a global mean of targets or across samples
}
\examples{
data(bulkATAC_raw_filtered, package = "compartmap")
atac_compartments <- getATACABsignal(filtered.data, allchrs = TRUE)
}
