% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChrs}
\alias{getChrs}
\title{Get the chromosomes from an object}
\usage{
getChrs(obj)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}
}
\value{
A character vector of chromosomes present in an object
}
\description{
Get the chromosomes from an object
}
\examples{
data("meth_array_450k_chr14", package = "compartmap")
getChrs(array.data.chr14)

}
