% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompartments.R
\name{getCompartments}
\alias{getCompartments}
\title{Run compartment inference}
\usage{
getCompartments(
  obj,
  res,
  chr,
  group,
  targets,
  bootstrap,
  num.bootstraps,
  genome,
  assay,
  boot.parallel,
  bpparams
)
}
\description{
Run compartment inference
}
