% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{Get the seqlengths of a chromosome}
\usage{
getSeqLengths(genome = c("hg19", "hg38", "mm9", "mm10"), chr = "chr14")
}
\arguments{
\item{genome}{The desired genome to use ("hg19", "hg38", "mm9", "mm10")}

\item{chr}{What chromosome to extract the seqlengths of}
}
\value{
The seqlengths of a specific chromosome
}
\description{
The goal for this function is to eliminate the need to lug around
large packages when we only want seqlengths for things.
}
\examples{
hg19.chr14.seqlengths <- getSeqLengths(genome = "hg19", chr = "chr14")
}
