% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTransform.R
\name{ifisherZ}
\alias{ifisherZ}
\title{Inverse Fisher's Z transformation}
\usage{
ifisherZ(zmat)
}
\arguments{
\item{zmat}{matrix of Fisher's Z transformed Pearson correlations or an
eignevector}
}
\value{
Back transformed Fisher's Z Pearson correlations
}
\description{
\code{ifisherZ} returns the inverse (squeezed) Fisher's Z transformed
Pearson's r.
}
\examples{

# Generate a random binary (-1, 1) matrix
(mat <- matrix(sample(c(1,-1), 25, replace = TRUE), ncol = 5))

# Correct matrix diag
diag(mat) <- 1

# Transform
(mat.transform <- fisherZ(mat))

#Back transform
ifisherZ(mat.transform)
}
