% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentCall.R
\name{is_open}
\alias{is_open}
\alias{is_closed}
\title{Find indices of open and closed compartments}
\usage{
is_open(x)

is_closed(x)
}
\arguments{
\item{x}{A \code{CompartmentCall} object}
}
\value{
\itemize{
\item \code{is_open}: boolean vector where TRUE means open
\item \code{is_closed}: boolean vector where TRUE means closed
}
}
\description{
Find indices of open and closed compartments
}
\concept{s7getters}
