% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSmoother.R
\name{meanSmoother}
\alias{meanSmoother}
\title{Windowed mean smoother}
\usage{
meanSmoother(x, k = 1, iter = 2, na.rm = TRUE, delta = 0)
}
\arguments{
\item{x}{Input data matrix: samples are columns, regions/loci are rows}

\item{k}{Number of windows to use (default k=1, i.e., 3 windows)}

\item{iter}{Number of iterations to smooth (default is 2)}

\item{na.rm}{Whether to remove NAs prior to smoothing (TRUE)}

\item{delta}{Convergence threshhold (overrides iter if > 0; default is 0)}
}
\value{
\preformatted{ Smoothed data matrix
}
}
\description{
TODO: farm out to C++ and test, at least when there are no NAs
}
\examples{
dummy <- matrix(rnorm(10000), ncol=25)
smooth.dummy <- meanSmoother(dummy)
smooth.dummy <- meanSmoother(dummy, iter=3) 
smooth.dummy <- meanSmoother(dummy, delta=1e-3) 

}
