% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSmoother.R
\name{meanSmoother}
\alias{meanSmoother}
\title{Windowed mean smoother}
\usage{
meanSmoother(x, k = 1, iter = 2, na.rm = TRUE)
}
\arguments{
\item{x}{Input data matrix where samples are columns and regions/loci are rows}

\item{k}{The number of windows to use (k=1 is 3 windows and k=2 is 5 windows)}

\item{iter}{The number of iterations to smooth}

\item{na.rm}{Whether to remove NAs prior to smoothing}
}
\value{
Smoothed data matrix
}
\description{
Windowed mean smoother
}
\examples{
dummy <- matrix(rnorm(10000), ncol=25)
smooth.dummy <- meanSmoother(dummy)

}
