% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_quantiles}
\alias{normalize_quantiles}
\title{Normalize quantiles
From limma version 3.64.3}
\usage{
normalize_quantiles(m, ties = TRUE)
}
\arguments{
\item{m}{Matrix of compartment call singular values}
}
\description{
Normalize quantiles
From limma version 3.64.3
}
