% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_quantiles}
\alias{normalize_quantiles}
\title{Normalize quantiles
From limma version 3.64.3}
\usage{
normalize_quantiles(A, ties = TRUE)
}
\arguments{
\item{A}{Numeric matrix of compartment call singular values}

\item{ties}{logical. If TRUE, ties in each column of A are treated in careful way.
tied values will be normalized to the mean of the corresponding pooled
quantiles.}
}
\description{
Normalize quantiles
From limma version 3.64.3
}
\keyword{internal}
