% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{plot.compartmap::CompartmentCall}
\alias{plot.compartmap::CompartmentCall}
\title{Plot singular values of a CompartmentCall object}
\usage{
\method{plot}{`compartmap::CompartmentCall`}(x)
}
\description{
Plot singular values of a CompartmentCall object
}
