% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialCompartments.R
\name{plot_dc}
\alias{plot_dc}
\title{Show differential compartments as overlaid \code{ggplot2::geom_rect()} over significant distances}
\usage{
plot_dc(
  ccall_pd,
  md,
  select,
  type = c("line", "bar"),
  show_md = TRUE,
  alpha_level = 0.05,
  fill = "maroon",
  alpha = 0.5,
  ylim = c(-0.5, 0.5),
  xlim = NULL,
  label_ids = TRUE
)
}
\arguments{
\item{ccall_pd}{The \verb{@df} slot of a \code{MultiCompartmapCall} object}

\item{md}{data.table output from \code{compute_mahalanobis()}}

\item{select}{Column name or index to compute differential compartments on.
Set this to plot the compartment scores of all input columns but show
differential compartments based only on the provided columns.}

\item{show_md}{Whether to plot the Mahalanobis distance}

\item{alpha_level}{Significance level to use (default: 0.05)}

\item{fill}{Color of the \code{geom_rect}}

\item{alpha}{Transparency of the \code{geom_rect}}

\item{ylim}{The y-axis limits}

\item{xlim}{The x-axis limits}

\item{label_ids}{Whether to differential bin ID labels}
}
\description{
Show differential compartments as overlaid \code{ggplot2::geom_rect()} over significant distances
}
\examples{
v <- c(1:4, 5, 7:9)
get_sequential_idx(v)
}
