% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompartmapCall.R
\name{plot_diff_compartments}
\alias{plot_diff_compartments}
\title{Compute differential compartments based on Mahalanobis distance}
\usage{
plot_diff_compartments(
  x,
  type = c("line", "bar"),
  cov_method = c("base", "robust", "mcd"),
  alpha_level = 0.05,
  show_md = TRUE,
  shade = "maroon",
  alpha = 0.5,
  ylim = c(-0.1, 0.1),
  xlim = NULL,
  label_ids = TRUE,
  select = NULL,
  color = NULL,
  fill = NULL,
  linewidth = 0.5
)
}
\arguments{
\item{x}{A \code{MultiCompartmentCall} object}

\item{cov_method}{Method to compute covariance. "base": \code{stats::cov},}

\item{alpha_level}{Significance level to use (default: 0.05)}

\item{show_md}{Whether to plot the Mahalanobis distance}

\item{shade}{Color of the \code{geom_rect}}

\item{alpha}{Transparency of the \code{geom_rect}}

\item{ylim}{The y-axis limits}

\item{xlim}{The x-axis limits}

\item{label_ids}{Whether to differential bin ID labels}

\item{select}{Column name or index to compute differential compartments on.
Set this to plot the compartment scores of all object columns but show
differential compartments based only on the provided columns.}

\item{color}{The line colors used for the each sample}

\item{fill}{The colors used for positive and negative values in bar plots}

\item{linewidth}{The width of lines in the line plot}
}
\description{
Compute differential compartments based on Mahalanobis distance
}
\concept{s7analysis}
