% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlobalMeans.R
\name{precomputeBootstrapMeans}
\alias{precomputeBootstrapMeans}
\title{Pre-compute the global means for bootstrapping compartments}
\usage{
precomputeBootstrapMeans(
  obj,
  targets = NULL,
  num.bootstraps = 100,
  assay = c("atac", "rna"),
  parallel = FALSE,
  num.cores = 1
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{targets}{Optional targets to shrink towards}

\item{num.bootstraps}{The number of bootstraps to compute}

\item{assay}{What type of assay the data are from}

\item{parallel}{Whether to run in parallel}

\item{num.cores}{How many cores to use for parallel processing}
}
\value{
A matrix of bootstrapped global means
}
\description{
Pre-compute the global means for bootstrapping compartments
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
scrna.bootstrap.global.means <- precomputeBootstrapMeans(k562_scrna_chr14, assay = "rna", num.bootstraps = 2)

}
