% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeEmptyBoots}
\alias{removeEmptyBoots}
\title{Remove bootstrap estimates that failed}
\usage{
removeEmptyBoots(obj)
}
\arguments{
\item{obj}{Input list object with elements 'pc' and 'gr'}
}
\value{
A filtered list object
}
\description{
This can happen if the correlation between the bins and eigenvector fails
theoretically we can recover these but need an additional utility to find
consensus
}
\keyword{internal}
