% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{resolution}
\alias{resolution}
\title{Get the resolution of the CompartmentCall}
\usage{
resolution(x)
}
\arguments{
\item{x}{A CompartmentCall object}
}
\description{
Get the resolution of the CompartmentCall
}
