% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCompartments.R
\name{scCompartments}
\alias{scCompartments}
\title{Estimate A/B compartments from single-cell sequencing data}
\usage{
scCompartments(
  obj,
  res = 1000000,
  parallel = FALSE,
  chr = NULL,
  targets = NULL,
  cores = 2,
  bootstrap = TRUE,
  num.bootstraps = 100,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  group = FALSE,
  assay = c("atac", "rna")
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{parallel}{Whether to run samples in parallel}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{targets}{Samples/cells to shrink towards}

\item{cores}{How many cores to use when running samples in parallel}

\item{bootstrap}{Whether we should perform bootstrapping of inferred compartments}

\item{num.bootstraps}{How many bootstraps to run}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{group}{Whether to treat this as a group set of samples}

\item{assay}{What type of single-cell assay is the input data ("atac" or "rna")}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{scCompartments} returns estimated A/B compartments from sc-seq data.
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
sc_compartments <- scCompartments(k562_scrna_chr14, parallel = FALSE, chr = "chr14", bootstrap = FALSE, genome = "hg19")
}
