% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCompartments.R
\name{scCompartments}
\alias{scCompartments}
\title{Estimate A/B compartments from single-cell RNA or ATAC sequencing data}
\usage{
scCompartments(
  obj,
  res = 1000000,
  chr = NULL,
  group = FALSE,
  targets = NULL,
  bootstrap = TRUE,
  num.bootstraps = 100,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  assay = c("atac", "rna"),
  boot.parallel = FALSE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{obj}{Input SummarizedExperiment object}

\item{res}{Compartment resolution in bp}

\item{chr}{What chromosome to work on (leave as NULL to run on all chromosomes)}

\item{group}{Whether to treat this as a group set of samples}

\item{targets}{Samples/cells to shrink towards}

\item{bootstrap}{Whether we should perform bootstrapping of inferred compartments}

\item{num.bootstraps}{How many bootstraps to run}

\item{genome}{What genome to work on ("hg19", "hg38", "mm9", "mm10")}

\item{assay}{What type of single-cell assay is the input data ("atac" or "rna")}

\item{boot.parallel}{Whether to run the bootstrapping in parallel. See details.}

\item{BPPARAM}{BiocParallelParam object to use for parallelization. See details.}
}
\value{
A RaggedExperiment of inferred compartments
}
\description{
\code{scCompartments} returns estimated A/B compartments from sc-seq data.
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
sc_compartments <- scCompartments(
  k562_scrna_chr14,
  chr = "chr14",
  parallel = FALSE,
  bootstrap = FALSE,
  genome = "hg19"
)
}
