% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkBins.R
\name{shrinkBins}
\alias{shrinkBins}
\title{Employ an eBayes shrinkage approach for bin-level estimates for A/B inference}
\usage{
shrinkBins(
  x,
  original.x,
  prior.means = NULL,
  chr = NULL,
  res = 1e+06,
  targets = NULL,
  jse = TRUE,
  assay = c("rna", "atac"),
  genome = c("hg19", "hg38", "mm9", "mm10")
)
}
\arguments{
\item{x}{Input SummarizedExperiment object}

\item{original.x}{Full sample set SummarizedExperiment object}

\item{prior.means}{The means of the bin-level prior distribution}

\item{chr}{The chromosome to operate on}

\item{res}{Resolution to perform the binning}

\item{targets}{The column/sample/cell names to shrink towards}

\item{jse}{Whether to use a James-Stein estimator (default is TRUE)}

\item{assay}{What assay type this is ("rna", "atac")}

\item{genome}{What genome are we working with ("hg19", "hg38", "mm9", "mm10")}
}
\value{
A list object to pass to getCorMatrix
}
\description{
\code{shrinkBins} returns shrunken bin-level estimates
}
\details{
This function computes shrunken bin-level estimates using a James-Stein estimator, reformulated as an eBayes procedure
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
shrunken.bin.scrna <- shrinkBins(x = k562_scrna_chr14,
                                 original.x = k562_scrna_chr14,
                                 chr = "chr14", assay = "rna")

}
