% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingPrep.R
\name{transformTFIDF}
\alias{transformTFIDF}
\title{Transform/normalize compartment calls using TF-IDF}
\usage{
transformTFIDF(obj, scale.factor = 1e+05)
}
\arguments{
\item{obj}{n x p input matrix (n = samples/cells; p = compartments)}

\item{scale.factor}{Scaling factor for the term-frequency (TF)}
}
\value{
A TF-IDF transformed matrix of the same dimensions as the input
}
\description{
Transform/normalize compartment calls using TF-IDF
}
\examples{

}
