% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTFIDF.R
\name{transformTFIDF}
\alias{transformTFIDF}
\title{Transform/normalize compartment calls using TF-IDF}
\usage{
transformTFIDF(obj, scale.factor = 100000)
}
\arguments{
\item{obj}{n x p input matrix (n = samples/cells; p = compartments)}

\item{scale.factor}{Scaling factor for the term-frequency (TF)}
}
\value{
A TF-IDF transformed matrix of the same dimensions as the input
}
\description{
Transform/normalize compartment calls using TF-IDF
}
\examples{

m <- 1000
n <- 100
mat <- round(matrix(runif(m*n), m, n))
#Input needs to be a tall matrix
tfidf <- transformTFIDF(mat)

}
