% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmap-S7.R
\name{unitarize}
\alias{unitarize}
\title{Unitarize the singular values in a CompartmentCall or MultiCompartmentCall}
\usage{
unitarize(x, ...)
}
\arguments{
\item{x}{A CompartmentCall object}
}
\description{
Unitarize the singular values in a CompartmentCall or MultiCompartmentCall
}
