% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifySE}
\alias{verifySE}
\title{Check if the assay is a SummarizedExperiment}
\usage{
verifySE(obj)
}
\arguments{
\item{obj}{Input object}
}
\description{
Check if the assay is a SummarizedExperiment
}
\examples{
data("k562_scrna_chr14", package = "compartmap")
verifySE(k562_scrna_chr14)
}
\keyword{internal}
